backend/WO/requirements.txt
appdirs==1.4.3
Django==1.10
django-reversion==2.0.8
packaging==16.8
psycopg2==2.7
pyparsing==2.2.0
six==1.10.0
toml==0.9.2
